LayoutTests/fast/forms/remove-form-inside-formdata-event-expected.txt

Tests that form submission gets cancelled when the form gets detached from inside the formdata event handler.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS frame.contentWindow.location.href.indexOf('test2.txt') != -1 is true
PASS successfullyParsed is true

TEST COMPLETE
